refactor sitemap test fixtures
